+++ packages/tnix-lsp/src/Server.spec.hs
@@ -42,6 +42,7 @@ spec = do
                 , "documentHighlightProvider" .= True
                 , "documentLinkProvider" .= object ["resolveProvider" .= False]
                 , "foldingRangeProvider" .= True
+                , "inlayHintProvider" .= object ["resolveProvider" .= False]
                 , "semanticTokensProvider"
                     .= object
                       [ "legend"